Fix build issue in amd64 Ubuntu.
(cherry picked from commit 3432a7f3656155f5f14d919089b1b5c1c5fd55d7)

+++ code/ROS/tank_agent/src/TankAgent.cpp
@@ -202,7 +202,7 @@ error_t readTankMsg(TankMsg *tankMsg) {
         return -1;
     }
 
-    size_t dataSize = std::min(tankMsg->dataLength, sizeof(tankMsg->data));
+    size_t dataSize = std::min((size_t) tankMsg->dataLength, sizeof(tankMsg->data));
     if (readSize(&tankMsg->data, dataSize)) {
         return -1;
     }
